Without withAppStaticProps
