+++ context/RouterContext.ts
@@ -11,7 +11,7 @@ export const RouterContext = createContext<RouterContextValue | null>(null);
 export const useRouterContext = () => {
   const context = useContext(RouterContext);
   if (!context) {
-    throw new Error('RouterContext not found. Did you forget to use "withAppStaticProps"?');
+    throw new Error("RouterContext not found. Did you forget to get app static props?");
   }
   return context;
 };
